refactor(http): move the REST endpoints into the plugin
The plugin now owns the whole REST surface: the game-facing groups and
their DTOs join the image routes in Moongate.Http.Plugin, consuming the
contracts from Moongate.Server.Abstractions. MoongateApiEndpointsPlugin
existed only to work around the old dependency cycle and is gone.
