+++ src/Moongate.Http.Plugin/Data/Api/LoginRequest.cs
@@ -1,4 +1,4 @@
-namespace Moongate.Server.Data.Api;
+namespace Moongate.Http.Plugin.Data.Api;
 
 /// <summary>Credentials posted to the login endpoint.</summary>
 public sealed record LoginRequest(string Username, string Password);
