+++ src/Moongate.Http.Plugin/Data/Api/PlayerMeResponse.cs
@@ -1,4 +1,4 @@
-namespace Moongate.Server.Data.Api;
+namespace Moongate.Http.Plugin.Data.Api;
 
 /// <summary>Who the caller's token says they are. Deliberately not their characters — see PlayerEndpoints.</summary>
 public sealed record PlayerMeResponse(string Username, string Level);
