+++ src/Moongate.Http.Plugin/Endpoints/AdminEndpoints.cs
@@ -3,12 +3,12 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Moongate.Http.Plugin.Interfaces;
 using Moongate.Http.Plugin.Services;
-using Moongate.Server.Data.Api;
+using Moongate.Http.Plugin.Data.Api;
 using Moongate.Server.Abstractions.Data.Config;
 using Moongate.Server.Abstractions.Interfaces.Accounts;
 using SquidStd.Core.Utils;
 
-namespace Moongate.Server.Endpoints;
+namespace Moongate.Http.Plugin.Endpoints;
 
 /// <summary>Staff-only shard administration.</summary>
 public sealed class AdminEndpoints : IApiEndpointRegistration
