+++ src/Moongate.Http.Plugin/Endpoints/PlayerEndpoints.cs
@@ -4,9 +4,9 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Moongate.Http.Plugin.Interfaces;
 using Moongate.Http.Plugin.Services;
-using Moongate.Server.Data.Api;
+using Moongate.Http.Plugin.Data.Api;
 
-namespace Moongate.Server.Endpoints;
+namespace Moongate.Http.Plugin.Endpoints;
 
 /// <summary>What an authenticated account may ask about itself.</summary>
 public sealed class PlayerEndpoints : IApiEndpointRegistration
