+++ src/Moongate.Server/Program.cs
@@ -10,7 +10,7 @@ using Moongate.Server.Autostart;
 using Moongate.Server.Abstractions.Data.Config;
 using Moongate.Server.Abstractions.Data.Events;
 using Moongate.Server.Data.Exceptions;
-using Moongate.Server.Endpoints;
+using Moongate.Http.Plugin.Endpoints;
 using Moongate.Server.Abstractions.Interfaces.Accounts;
 using Moongate.Server.Abstractions.Interfaces.Items;
 using Moongate.Server.Abstractions.Interfaces.Mobiles;
@@ -114,7 +114,6 @@ await ConsoleApp.RunAsync(
                 if (!disableWebPlugin)
                 {
                     builder.Add<MoongateHttpPlugin>();
-                    builder.Add<MoongateApiEndpointsPlugin>();
                 }
                 else
                 {
