+++ tests/Moongate.Tests/Http/CharacterAdminEndpointsTests.cs
@@ -6,8 +6,8 @@ using Moongate.Http.Plugin.Data;
 using Moongate.Network.Packets.Incoming;
 using Moongate.Network.Types;
 using Moongate.Persistence.Entities;
-using Moongate.Server.Data.Api;
-using Moongate.Server.Endpoints;
+using Moongate.Http.Plugin.Data.Api;
+using Moongate.Http.Plugin.Endpoints;
 using Moongate.Server.Abstractions.Interfaces.Accounts;
 using Moongate.Server.Services.Accounts;
 using Moongate.Tests.Support;
